Backend/requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.7
django-filter==22.1
django-versatileimagefield==3.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
drf-spectacular==0.25.1
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
packaging==23.0
Pillow==9.4.0
pymongo==4.3.3
pyrsistent==0.19.3
python-magic==0.4.27
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.2.4
swagger-spec-validator==3.0.3
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14
